Configuration improvements
Add a 'slim' section to config.php.dist so that we can configure
the mode and other configuration more easily. Ensure that the
view is instantiated in index.php though.

Also, only set error_reporting and display_errors when the mode is
set to development.

Note that we load config.php.dist if a config.php is not set. This
mean that it is assumed that the live site has its own config.php.