Consolidate Slug handling to Model\API\Event
Slug handling was handled in Model\API\Event and Model\Event. This
commit consolidates to just Model\API\Event as I think that the Event
model should not be database aware.

Arguably, Model\API\Event shouldn't be database aware either and there
should be a Service\Event class for this. That's left for another day.