Remove superfluous and misleading echo's
In the each call to render seems to be prefixed with the echo keyword
but the render method does not return data; it adds the rendered output
to the view cache / output buffer so that it can be echo'd later
by Slim itself.

In this commit I remove all superfluous echo's since it may mislead
people into thinking that the echo's actually have a function and use
it to circumvent the view mechanism of Slim or have unpredictable bugs.