Added a Composer file with license
With this commit we add the composer.json and lock file together
with a license file indicating that this project is under the
MIT license (discussed with @akrabat to be sure).

In addition to the above I have prevented the .git folder in the
vendor folder from being committed.

The vendor folder itself will be submitted in a separate commit
to make it easier for a code reviewer to see which changes are
the onces actually needing a review.