Return headers after a POST call
When we submit an event we want to redirect to the new
event page. This is only possible if we have access to the
Location header (that contains the new event's API url).

This commit adds support for the apiPost method to return
header information, which in turn can be consumed by the
caller to retrieve said location header from.