Submit new events to the API
In this commit I add support for submitting a new event to the API
using a data array that contains all fields as supported by the
POST method for the `/v2.1/events` endpoint (see the API documentation
for more information).

In addition a field timezone is supported that takes a complete
timezone (such as Europe/Amsterdam) and populates the tz_continent
and tz_place fields.