Add form to submit new Events
In order to be able to submit events to the API we need a
form to take the event properties, validate those, send them
to the API and redirect to the details page for that event
if this all succeeds.

In this commit I have

- added the form definition and template
- changed the event controller to construct the form instance
- changed the event controller to validate the form
- changed the event controller to save the event to the API and
- redirect if the event was submitted correctly to its detail page

Some styling was added so that the event form looks passable; it
is recommended to review the designs at a later stage.