Enable Twig Cache
Because Twig becomes increasingly slower when you load a large object, and
a form is a large object, have I enabled the cache for all environments except
development.

The default caching folder is set to `/tmp/joindin-twig-cache/[environment]`
but can be changed in the configuration file. The cache may even be turned off
by providing false.

For the development environment the cache setting is always set to false; if
you want to test with cache enabled you need to edit the index.php and comment
the disabling of the cache.