Retrieve host username in event collection
After creating the event entity object for each event in the collection
we need to find the username of each host so that we can link to them.

This is currently used in the pending events list page, but due to the
caching in redis, is not an especially expensive operation to do anyway.