Use the API's "all" filter for the events list
If the page query parameter is not set, then do not pass a start number
to the API's so that the event's "all" filter will return the first
page of events where at least one event is in the future. That is, we
get a list of events positioned at "now".

In this case, we have to work out the current page number ourselves,
which we do by looking at the this_page meta field which contains the
correct start parameter.