Use start parameter for event list page
As the event list page uses the API's `all` collection, we can't use the
page parameter in the query any more as the start parameter is no longer
on a page boundary. Hence we expose the start parameter directly in the
query string.

Note that this means that if a user goes all the way back to the very
start of the collection and then goes forward again, the pagination
boundary will change as we can't have a negative start number.