Add event nav-bar to Edit event page
Place the nav-bar under the edit event title and update the styling so
that we don't have two borders next to each other with no text between
them.

Add the correct Bootstrap grid tags so that the alignment is better.
Also remove the test that we have a user object as we can't edit an
event unless we're logged in.