Move admin nav bar on Events page to below title
For consistency with the nav-bar on the event's detail pages, move the
(admin only) nav-bar on the events list page to under the title too.

When the nav-bar is displayed, wrap in a div with the 'page-header'
class so that the bottom border of the nav-bar is displayed. Note that
as the nav-bar is admin only and we don't want a double border when
there is no nav-bar, we have to render the 'page-header' div within the
check that the user is an admin. Otherwise, we just render the title.