Fixes possibly endles while-loop in comment-field
In some edge-cases it was possible that the while-loop that adapted the
height of the comment field to the contents height went berserk and
extended the height indefinitely. This commit changes that loop to one
single function call so it will not be able to run indefinitely and
leave the browser unresponsive.

Thanks to Elizabeth Smith for finding the issue and testing the fix.

Fixes JOINDIN-713