Adds QR-Code links to event- and talk-details
This commit adds a link to an QR-Code to the event- and the
talk-details. That way organizers and speakers can easily add a QR-Code
to their slides for easier reference of the event or talk. That makes it
easier for attendees to find the talk or event and rate it.