Offers to resend welcome email to un verified user upon login attempt
If a user that is not verified tries to login then this change will show a flash message to that effect and provide a link which the user can click in order to have the validation email resent, this does not resend the email automatically, rather the user needs to enter their email address.  I personally think this is a more secure way to go about it rather than having a single link click send the email.