Added the CTA to submit an event to the homepage
I was talking with Mark Jansen of the WordPress Meetup Enschede. He is not a very regular user of joind.in and he mentioned that he missed a CTA to add an event on the homepage. I decided it was a good idea to get less experienced used on board and create a PR. I hope you like this ;)