Add honeypot for contact form
This adds a Form-Field that has to be empty. A so called HoneyPot. A
trap for bots that usually fill out every form field that seems
relevant. As the introduced field needs to be empty every
form-submission that contains that field will automatically be filtered
out as the form wasn't valid.