Optimize Platform.sh-logo in footer
The Platform.sh logo was a bit on the huge size in the footer-line. This commit fixes that by removing the fixed width and height that did not fit and replacing it with (currently hard-coded) style-information to make it more along the line of what the original CI of platform wanted it to look like.